Refactored View actors moving common parts to FXViewActor
