+++ client/src/main/scala/it/cwmp/client/view/authentication/AuthenticationViewActor.scala
@@ -1,68 +1,28 @@
 package it.cwmp.client.view.authentication
 
-import akka.actor.{Actor, ActorRef}
 import it.cwmp.client.controller.messages.AuthenticationRequests.{LogIn, SignUp}
-import it.cwmp.client.controller.messages.Initialize
-import it.cwmp.client.controller.{ActorAlertManagement, ActorViewVisibilityManagement}
-import javafx.application.Platform
-import javafx.embed.swing.JFXPanel
+import it.cwmp.client.view.FXViewActor
 
 /**
   * Actor assigned to the management of the display of the authentication screen and of the events generated by it.
   *
   * @author Elia Di Pasquale
   */
-case class AuthenticationViewActor() extends Actor
-  with ActorAlertManagement with ActorViewVisibilityManagement {
+case class AuthenticationViewActor() extends FXViewActor {
 
-  var fxController: AuthenticationFXController = _
-
-  var controllerActor: ActorRef = _
+  override protected var fxController: AuthenticationFXController = _
 
   override def preStart(): Unit = {
     super.preStart()
+    fxController = AuthenticationFXController(new AuthenticationStrategy {
+      override def performLogIn(username: String, password: String): Unit =
+        controllerActor ! LogIn(username, password)
 
-    new JFXPanel // initializes JavaFX
-    Platform setImplicitExit false
-    runOnUIThread(() => {
-      fxController = AuthenticationFXController(new AuthenticationStrategy {
-        override def performLogIn(username: String, password: String): Unit =
-          controllerActor ! LogIn(username, password)
+      override def performPasswordCheck(password: String, confirmPassword: String): Boolean =
+        password == confirmPassword
 
-        override def performPasswordCheck(password: String, confirmPassword: String): Boolean =
-          password == confirmPassword
-
-        override def performSignUp(username: String, password: String): Unit =
-          controllerActor ! SignUp(username, password)
-      })
+      override def performSignUp(username: String, password: String): Unit =
+        controllerActor ! SignUp(username, password)
     })
   }
-
-  override def receive: Receive = alertBehaviour orElse visibilityBehaviour orElse {
-    case Initialize => controllerActor = sender()
-  }
-
-  override protected def onErrorAlertReceived(title: String, message: String): Unit = {
-    super.onErrorAlertReceived(title, message)
-    onAlertReceived()
-  }
-
-  override protected def onInfoAlertReceived(title: String, message: String): Unit = {
-    super.onInfoAlertReceived(title, message)
-    onAlertReceived()
-  }
-
-  /**
-    * When receiving an alert should enable buttons and hide loading
-    */
-  private def onAlertReceived(): Unit = {
-    fxController enableViewComponents()
-    fxController hideLoading()
-  }
-
-  override protected def onHideGUI(): Unit = {
-    super.onHideGUI()
-    fxController hideLoading()
-  }
-
 }
